WIP: Use local storage to save user login session
Instead of session storage, use local storage to save user session.
This way it is possible to open up pages to new tabs without the need
of logging the user in again.

Add an event listener to see if the user has been logged out from any
of the active tabs and log out the user from all of the remaining tabs.

TODO: We need to clear the local storage when the user closes the
browser without logging out first